feat : 객관식 채점 생성

+++ src/main/java/com/example/interviewgod/domain/question/dto/request/SubmitAnswerRequest.java
@@ -3,5 +3,5 @@ package com.example.interviewgod.domain.question.dto.request;
 
 import java.util.List;
 
-public record SubmitAnswerRequest(Long selfIntroduceId, Integer sessionId, List<Integer> answer) {
+public record SubmitAnswerRequest(Long selfIntroduceId, Long sessionId, List<Integer> answer) {
 }
